+++ cli/jitp/createDeviceCertificate.ts
