+++ export.ts
@@ -2,5 +2,5 @@ export * from './cli/jitp/caFileLocations'
 export * from './cli/jitp/certsDir'
 export * from './cli/jitp/createCA'
 export * from './cli/jitp/deviceFileLocations'
-export * from './cli/jitp/generateDeviceCertificate'
+export * from './cli/jitp/createDeviceCertificate'
 export * from './cdk/helper/getIotEndpoint'
